Created DockerImageGarbageCollection
