+++ docker-api/src/main/resources/configdefinitions/docker.def
@@ -11,3 +11,6 @@ maxPerRouteConnections int default = 10
 maxTotalConnections int default = 100
 connectTimeoutMillis int default = 100000 # 100 sec
 readTimeoutMillis int default = 1800000 # 30 min
+
+imageGCEnabled bool default = true
+imageGCMinTimeToLiveMinutes int default = 120
